+++ templates/hello-world/controllers/common.js
@@ -3,13 +3,11 @@ exports.setRoutes = function (next) {
         route = NA.webconfig.routes;
 
     route.unshift({
-        "key": "add_fr-fr",
         "url": "/ajoutee/",
 		"output": "/ajoutee.html",
 		"view": "content.htm"
     });
     route.unshift({
-        "key": "add_en-us",
         "url": "/english/added/",
 		"output": "/english/added.html",
 		"view": "content.htm",
@@ -17,17 +15,4 @@ exports.setRoutes = function (next) {
     });
 
     next(); 
-};
-
-exports.changeDom = function (next, locals) {
-	var NA = this,
-        $ = locals.virtualDom();
-
-    for (let route of NA.webconfig.routes) {
-    	if (route.key === "home_" + locals.common.menu.href) {
-			$("div > a").attr("href", route.url.slice(1));
-    	}
-    }
-
-    next($);
 };