+++ templates/hello-world/controllers/index.js
@@ -1,5 +1,5 @@
 exports.changeVariations = function (next, variations) {
-	variations.specific.helloWorld = variations.specific.helloWorld + "!!!!!!!!!!!!!!!!!!!!";
+	variations.specific.helloWorld = variations.specific.helloWorld + "!";
 
 	next();
 };